+++ pages/admin/index.vue
@@ -12,7 +12,13 @@ const totalRevenue = ref(0);
 const reservations = ref([]);
 const todayAppointments = ref(0);
 
+const loadingReservations = ref(true);
+const loadingPayments = ref(true);
+const loadingRevenue = ref(true);
+const loadingPatients = ref(true);
+
 const fetchReservations = async () => {
+  loadingReservations.value = true;
   try {
     const response = await fetch('/api/reservations', {
       method: 'GET',
@@ -29,11 +35,14 @@ const fetchReservations = async () => {
     ).length; // Count today's appointments
   } catch (err) {
     console.error('Error fetching reservations:', err);
+  } finally {
+    loadingReservations.value = false
   }
 };
 
 
 const fetchPayments = async () => {
+  loadingPayments.value = true;
   try {
     const response = await fetch('/api/admin/payment', {
       method: 'GET',
@@ -43,39 +52,47 @@ const fetchPayments = async () => {
     totalRevenue.value = payment.value.reduce((sum, p) => sum + p.totalAmount, 0);
   } catch (error) {
     console.error('Error fetching payments:', error);
+  } finally {
+    loadingPayments.value = false; // Stop loading
   }
 };
 
 // Delete a payment
 const deletePayment = async (id) => {
-  try {
-    const response = await fetch(`/api/payment/${id}`, {
-      method: 'DELETE',
-    });
-    if (response.ok) {
-      payment.value = payment.value.filter((p) => p.id !== id);
-      totalRevenue.value = payment.value.reduce((sum, p) => sum + p.totalAmount, 0);
-    } else {
-      console.error('Error deleting payment');
+  const confirmed = window.confirm('คุณแน่ใจหรือว่าต้องการลบการชำระเงินนี้?'); // Confirmation message
+  if (confirmed) {
+    try {
+      const response = await fetch(`/api/payment/${id}`, {
+        method: 'DELETE',
+      });
+      if (response.ok) {
+        payment.value = payment.value.filter((p) => p.id !== id);
+        totalRevenue.value = payment.value.reduce((sum, p) => sum + p.totalAmount, 0);
+      } else {
+        console.error('Error deleting payment');
+      }
+    } catch (error) {
+      console.error('Error deleting payment:', error);
     }
-  } catch (error) {
-    console.error('Error deleting payment:', error);
   }
 };
 
 const deleteReservation = async (id) => {
-  try {
-    const response = await fetch(`/api/reservations?id=${id}`, {
-      method: 'DELETE',
-    });
-    if (!response.ok) {
-      throw new Error(`Error deleting reservations: ${response.statusText}`);
+  const confirmed = window.confirm('คุณแน่ใจหรือว่าต้องการลบนัดหมายนี้?'); // Confirmation message
+  if (confirmed) {
+    try {
+      const response = await fetch(`/api/reservations?id=${id}`, {
+        method: 'DELETE',
+      });
+      if (!response.ok) {
+        throw new Error(`Error deleting reservations: ${response.statusText}`);
+      }
+      await response.json();
+      // Optionally, remove the deleted reservation from the reservations list
+      reservations.value = reservations.value.filter((res) => res.id !== id);
+    } catch (error) {
+      console.error(error.message);
     }
-    await response.json();
-    // Optionally, remove the deleted reservation from the reservations list
-    reservations.value = reservations.value.filter((res) => res.id !== id);
-  } catch (error) {
-    console.error(error.message);
   }
 };
 
@@ -89,10 +106,17 @@ definePageMeta({
 });
 
 onMounted(async () => {
+  loadingPatients.value = true; // Start loading
   await userStore.fetchUsers().then(() => {
     totalPatients.value = userStore.users.length;
+    loadingPatients.value = false; // Finished loading
   });
-  await fetchPayments();
+
+  loadingRevenue.value = true; // Start loading
+  await fetchPayments().then(() => {
+    loadingRevenue.value = false; // Finished loading
+  });
+
   await fetchReservations(); // Ensure reservations and todayAppointments are fetched correctly
 });
 </script>
@@ -109,48 +133,56 @@ onMounted(async () => {
           <h2 class="card-title text-white font-light text-3xl pl-5 pt-5">รายได้ทั้งหมด</h2>
           <div class="flex mt-8">
             <div class="flex ml-5 w-full mt-8">
-              <p class="text-white text-4xl mt-1">{{ totalRevenue }} บาท</p>
+              <p class="text-white text-4xl mt-1" :class="{ 'text-xl': loadingRevenue }">
+                {{ loadingRevenue ? 'กำลังโหลด...' : totalRevenue + ' บาท' }}
+              </p>
             </div>
-            <div class=" absolute bottom-3 right-1 justify-end mr-5">
+            <div class="absolute bottom-3 right-1 justify-end mr-5">
               <img src="https://img2.pic.in.th/pic/users_391194.png" alt="logo_user" class="w-[75px]">
             </div>
           </div>
         </div>
+
         <div class="card bg-[#00AC97] w-full h-44 shadow-xl transform-transition bg-opacity-65">
           <h2 class="card-title text-white font-light text-3xl pl-5 pt-5">นัดหมายวันนี้</h2>
           <div class="flex mt-8">
             <div class="flex-1 ml-5 mt-8">
-              <p class="text-white text-4xl mt-2">{{ todayAppointments }} คน</p>
+              <p class="text-white text-4xl mt-2" :class="{ 'text-xl': loadingReservations }">
+                {{ loadingReservations ? 'กำลังโหลด...' : todayAppointments + ' คน' }}
+              </p>
             </div>
             <div class="flex-1 flex justify-end mr-5">
               <img src="https://img2.pic.in.th/pic/qc_8790314-1.png" alt="logo_user">
             </div>
           </div>
         </div>
-        <!-- This section displays today's appointments count -->
+
         <div class="card bg-[#515262] w-full h-44 shadow-xl transform-transition bg-opacity-65">
           <h2 class="card-title text-white font-light text-2xl pl-5 pt-5">รวมผู้ป่วยทั้งหมด</h2>
           <div class="flex mt-8">
             <div class="flex-1 ml-5 mt-8">
-              <p class="text-white text-4xl mt-2">{{ totalPatients }} คน</p>
-              <!-- Correctly bound to todayAppointments -->
+              <p class="text-white text-4xl mt-2" :class="{ 'text-xl': loadingPatients }">
+                {{ loadingPatients ? 'กำลังโหลด...' : totalPatients + ' คน' }}
+              </p>
             </div>
             <div class="flex-1 flex justify-end mr-5">
               <img src="https://img2.pic.in.th/pic/job_10485121.png" alt="logo_user">
             </div>
           </div>
         </div>
-
       </div>
     </div>
+
     <div class="bg-white rounded-lg h-[400px] shadow-lg">
       <div class="bg-[#FF8128] rounded-t-lg h-14 flex justify-center items-center bg-opacity-95 shadow-sm">
-        <h1 class="text-white text-3xl">
-          นัดหมายวันนี้
-        </h1>
+        <h1 class="text-white text-3xl">นัดหมายวันนี้</h1>
       </div>
       <div class="grid sm:grid-cols-4 grid-cols-1 gap-3 p-4">
-        <div v-for="(reservation, index) in reservations" :key="reservation.id" class="bg-base-100 rounded-xl mb-5 border-2 border-accent shadow-md">
+        <div v-if="loadingReservations" class="col-span-1 flex items-center justify-center">
+          <p class="text-gray-500">กำลังโหลดข้อมูล...</p>
+        </div>
+        <div v-for="(reservation, index) in reservations" :key="reservation.id"
+          class="bg-base-100 rounded-xl mb-5 border-2 border-accent shadow-md">
           <div class="flex justify-center mt-4">
             <img src="https://img2.pic.in.th/pic/volunteer_11077481.png" alt="logo-user" class="w-16">
           </div>
@@ -158,7 +190,6 @@ onMounted(async () => {
             <h1 class="text-center">ข้อมูลนัดหมาย</h1>
           </div>
           <div class="ml-4 mt-3 sm:text-sm text-[10px]">
-            
             <h1 class="text-base">{{ reservation.email }}</h1>
             <h1 class="text-base">ชื่อ : {{ reservation.firstname }} {{ reservation.lastname }}</h1>
             <h1 class="text-base">วันนัด : {{ formatDate(reservation.date) }}</h1>
@@ -166,8 +197,10 @@ onMounted(async () => {
             <h1 class="text-base">บัตรประชาชน : {{ reservation.cdnumber || 'N/A' }}</h1>
 
             <div class="flex gap-2 justify-center mb-2 mt-2">
-              <button @click="deleteReservation(reservation.id)" class="btn sm:w-28 w-16 shadow-md bg-red-500 text-white">
-                <Trash /> <p class="font-light hidden sm:block">ลบ</p>
+              <button @click="deleteReservation(reservation.id)"
+                class="btn sm:w-28 w-16 shadow-md bg-red-500 text-white">
+                <Trash />
+                <p class="font-light hidden sm:block">ลบ</p>
               </button>
               <button class="btn btn-accent sm:w-28 w-16">
                 <div class="flex">
@@ -179,18 +212,17 @@ onMounted(async () => {
           </div>
         </div>
       </div>
-
     </div>
 
     <div class="bg-white rounded-lg h-96 mt-5">
       <div class="bg-[#FF8128] rounded-t-lg h-14 flex justify-center items-center bg-opacity-95">
-        <h1 class="text-white text-3xl">
-          รายการล่าสุด
-        </h1>
+        <h1 class="text-white text-3xl">รายการล่าสุด</h1>
       </div>
       <div class="overflow-x-auto mt-3">
-        <table class="table">
-          <!-- head -->
+        <div v-if="loadingPayments" class="flex justify-center p-4">
+          <p class="text-gray-500">กำลังโหลดข้อมูล...</p>
+        </div>
+        <table v-else class="table">
           <thead>
             <tr>
               <th>
@@ -200,14 +232,17 @@ onMounted(async () => {
                 <p class="text-center">ชื่อลูกค้า</p>
               </th>
               <th>
-                <p class="text-center">ชื่อสินค้า</p>
+                <p class="text-center">รหัสสินค้า</p>
+              </th>
+              <th>
+                <p class="text-center">ราคา</p>
               </th>
               <th>
                 <p class="text-center">จ่ายแบบ</p>
               </th>
               <th>
                 <p class="text-center">สถานะ</p>
-              </th> <!-- Column for Status -->
+              </th>
               <th></th>
             </tr>
           </thead>
@@ -226,52 +261,39 @@ onMounted(async () => {
                   </div>
                   <div>
                     <div class="font-light">
-              <td>{{ payment.diagnosis?.patient?.firstname }} {{ payment.diagnosis?.patient?.lastname }}</td>
+                      {{ payment.diagnosis?.patient?.firstname }} {{ payment.diagnosis?.patient?.lastname }}
+                    </div>
+                  </div>
+                </div>
+              </td>
+              <td>
+                <p class="text-center">{{ payment.orderNumber }}</p>
+              </td>
+              <td>
+                <p class="text-center">{{ payment.totalAmount }} บาท</p>
+              </td>
+              <td>
+                <p class="text-center">พร้อมเพย์</p>
+              </td>
+              <td>
+                <p class="text-center">{{ payment.status }}</p>
+              </td>
+              <th>
+                <div class="flex gap-2">
+                  <button @click="deletePayment(payment.id)" class="btn btn-md btn-danger font-light">
+                    <Trash />
+                  </button>
+                  <button class="btn btn-md btn-accent font-light text-white">เพิ่มเติม</button>
+                </div>
+              </th>
+            </tr>
+          </tbody>
+        </table>
       </div>
     </div>
-    </div>
-    </td>
-    <td>
-      <p class="text-center">{{ payment.orderNumber }}</p>
-    </td>
-    <td>
-      <p class="text-center">พร้อมเพย์</p>
-    </td>
-    <td>
-      <p class="text-center">{{ payment.status }}</p>
-    </td>
-    <th>
-      <div class="flex gap-2">
-        <button @click="deletePayment(payment.id)" class="btn btn-md btn-danger font-light">
-          <Trash />
-        </button>
-        <button class="btn btn-md btn-accent font-light text-white">เพิ่มเติม</button>
-      </div>
-    </th>
-    </tr>
-    </tbody>
-    <!-- foot -->
-    </table>
-    </div>
-    </div>
-
-    <!-- <div class="flex gap-3">
-      <div class="flex-1">
-        <div class=" p-4 bg-white mt-5 shadow-md rounded-md">
-          <div class="flex justify-center">
-            <h1 class="text-3xl font-light">สรุปรายได้</h1>
-          </div>
-          <div class="mt-5">
-            <ApexBarChart />
-          </div>
-        </div>
-      </div>
-      <div class="flex-1">
-
-      </div>
-    </div> -->
   </adminLayouts>
 </template>
+
 
 <style scoped>
 .card {
